+++ redfrontier/lib/constants.dart
@@ -1,0 +1,3 @@
+class ProjectConstants {
+  static String recaptchaKey = "75078611-2E67-4365-A0C0-582B2702063F";
+}