+++ redfrontier/lib/screens/news_screen.dart
@@ -28,19 +28,25 @@ class _NewsScreenState extends State<NewsScreen> {
   Widget build(BuildContext context) {
     final uid = gpc.read(currentRFUserProvider)!.id;
     return Scaffold(
-      body: FutureBuilder(
-        future: FirestoreReportService.getMyReports(uid),
+      body: StreamBuilder(
+        stream: FirestoreReportService.getFeedReportsAsStream(uid),
         builder: (context, snapshot) {
-          if (snapshot.connectionState == ConnectionState.done) {
-            final data = snapshot.data;
-            if (data == null || data.isEmpty)
-              return Text('No Reports').center();
+          if (snapshot.hasData) {
+            final reports = (snapshot.data!.docs)
+                .map((e) => e.data())
+                .map((e) => Report.fromMap(e as Map<String, dynamic>))
+                .toList();
+
+            if (reports.isEmpty) {
+              return Text('No Reports to Display');
+            }
             return ListView.builder(
-              itemCount: data.length,
-              itemBuilder: (context, index) => ReportBody(model: data[index]),
+              itemCount: reports.length,
+              itemBuilder: (context, index) =>
+                  ReportBody(model: reports[index]),
             );
           }
-          return CircularProgressIndicator().center();
+          return Text('Unable to Fetch');
         },
       ),
     );