+++ redfrontier/lib/services/storagr/firebase_storage.dart
@@ -1,10 +1,14 @@
+import 'dart:async';
 import 'dart:io';
 
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:image_picker/image_picker.dart';
+import 'package:redfrontier/common/dialogs.dart';
+import 'package:redfrontier/main.dart';
 
 class FirebaseStorageService {
   static Future<String?> uploadImage(File file) async {
+    final Completer<String?> completer = Completer<String>();
     try {
       String fileName = DateTime.now().millisecondsSinceEpoch.toString();
       Reference storageRef =
@@ -14,10 +18,18 @@ class FirebaseStorageService {
         // Get the download URL after the image is uploaded
         String imageUrl = await storageRef.getDownloadURL();
         print('Image uploaded. URL: $imageUrl');
-        return imageUrl;
+        completer.complete(imageUrl);
       });
     } catch (e) {
       print('Error uploading image: $e');
+      return CustomDialogs.showDefaultAlertDialog(
+        navigatorKey.currentState!.context,
+        contentTitle: 'ImageUploadError',
+        contentText: 'unable to upload image',
+      );
+      completer.complete(null);
     }
+
+    return completer.future;
   }
 }